Removes unused components
Clean up progress functions

Add legend back in
