+++ src/components/app.css
@@ -1,11 +1,3 @@
-/* #app {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-} */
 .App {
   text-align: center;
   background-color: white;
