+++ src/components/app.tsx
@@ -1,9 +1,8 @@
-import React, { useEffect, useState } from "react";
-import { Provider } from "./ui/provider";
-import { TabContainer } from "./tabs/tab-container";
+import React, { useEffect } from "react";
 import { initializePlugin } from "@concord-consortium/codap-plugin-api";
 import { isNonEmbedded } from "../utils/embed-check";
-// import { ProgressOverlay } from "./ui/progress-overlay";
+import { TabContainer } from "./tabs/tab-container";
+import { Provider } from "./ui/provider";
 
 import "./app.css";
 
@@ -11,13 +10,10 @@ const kPluginName = "NASA Earth Data Plugin";
 const kVersion = "0.0.1";
 const kInitialDimensions = {
   width: 315,
-  height: 400
+  height: 410
 };
 
 export const App = () => {
-  const [progress, setProgress] = useState({ current: 0, total: 0 });
-  const [showProgress, setShowProgress] = useState(false);
-
   useEffect(() => {
     if (isNonEmbedded()) {
       return; // Skip initialization if noEmbed parameter exists
@@ -28,17 +24,7 @@ export const App = () => {
 
   return (
     <Provider>
-      <TabContainer current={progress.current} total={progress.total} isVisible={showProgress}
-          progressCallback={(current, total) => {
-          setProgress({ current, total });
-          setShowProgress(current > 0 && current < total);
-        }}
-      />
-      {/* <ProgressOverlay
-        current={progress.current}
-        total={progress.total}
-        isVisible={showProgress}
-      /> */}
+      <TabContainer />
     </Provider>
   );
 };
