+++ src/components/tabs/dataset-tab.scss
@@ -1,3 +1,10 @@
+.legend-container{
+  margin: 8px 0;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+}
+
 .divider {
   width: 295px;
   border: solid 1px #979797;
