+++ src/components/tabs/dataset-tab.tsx
@@ -3,28 +3,26 @@ import {
   addDataContextChangeListener,
   ClientNotification
 } from "@concord-consortium/codap-plugin-api";
-import { DatasetSelector } from "../dataset-selector/dataset-selector";
+import { DataManager, kDataContextName } from "../../models/data-manager";
 import { kNeoDatasets } from "../../models/neo-datasets";
 import { isNonEmbedded } from "../../utils/embed-check";
-import { DataManager, kDataContextName, ProgressCallback } from "../../models/data-manager";
+import { DatasetSelector } from "../dataset-selector/dataset-selector";
+import { ProgressContainer } from "./progress-container";
 
 import "./dataset-tab.scss";
-import { Box, Stack } from "@chakra-ui/react";
-interface DatasetTabProps {
-  current: number;
-  total: number;
-  isVisible: boolean;
-  progressCallback: ProgressCallback;
-}
 
-export const DatasetTab: React.FC<DatasetTabProps> = ({ current, total, isVisible, progressCallback }) => {
+export const DatasetTab: React.FC = () => {
   const [, setListenerNotification] = useState<string>();
   const defaultNeoDatasetId = kNeoDatasets[0].id;
   const [selectedNeoDatasetId, setSelectedNeoDatasetId] = useState<string>(defaultNeoDatasetId);
-  const [isFetching, setIsFetching] = useState(false);
+  const [progress, setProgress] = useState({ current: 0, total: 0 });
+  const [showProgress, setShowProgress] = useState(false);
   const [dataManager] = useState(() => {
     const manager = new DataManager();
-    manager.setProgressCallback(progressCallback);
+    manager.setProgressCallback((current, total) => {
+      setProgress({ current, total });
+      setShowProgress(current > 0 && current < total);
+    });
     return manager;
   });
 
@@ -49,9 +47,8 @@ export const DatasetTab: React.FC<DatasetTabProps> = ({ current, total, isVisibl
 
   const handleGetData = async () => {
     if (selectedNeoDataset) {
-      setIsFetching(true);
+      setShowProgress(true);
       await dataManager.getData(selectedNeoDataset);
-      setIsFetching(false);
     }
   };
 
@@ -59,13 +56,14 @@ export const DatasetTab: React.FC<DatasetTabProps> = ({ current, total, isVisibl
     <div className="App">
       <h1>NASA Earth Data</h1>
       <DatasetSelector selectedDataset={selectedNeoDatasetId} onDatasetChange={handleDatasetChange} />
-      <img src={selectedNeoDataset?.legendImage} alt={`${selectedNeoDataset?.label} legend`} />
-
+      <div className="legend-container">
+        <img src={selectedNeoDataset?.legendImage} alt={`${selectedNeoDataset?.label} legend`} />
+      </div>
       <div className="divider" />
       <div className="footer">
-        {isVisible && <ProgressContainer current={current} total={total}/>}
+        {showProgress && <ProgressContainer current={progress.current} total={progress.total}/>}
         <div className="footer-buttons-container">
-          <button className="get-data-button" disabled={isFetching} onClick={handleGetData}
+          <button className="get-data-button" disabled={showProgress} onClick={handleGetData}
             title="Fetch data from NASA and send to CODAP">
               Get Data
           </button>
@@ -73,24 +71,4 @@ export const DatasetTab: React.FC<DatasetTabProps> = ({ current, total, isVisibl
       </div>
     </div>
   );
-};
-
-interface IProgressContainerProps {
-  current: number;
-  total: number;
-}
-
-const ProgressContainer: React.FC<IProgressContainerProps> = ({current, total}) => {
-  const percentage = Math.round((current / total) * 100);
-
-  return (
-    <Stack direction="column" gap={2} className="progress-container">
-        <Box width="150px" fontSize="xs" color="gray.600" textAlign="center">
-          Loading images: {current}/{total}
-        </Box>
-        <Box w="100%" h="2" bgColor="gray.100" borderRadius="full" overflow="hidden">
-          <Box w={`${percentage}%`} h="100%" bgColor="blue.500" transition="width 0.2s"
-            borderRadius="full"/>
-        </Box>
-    </Stack>);
 };
